# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Barber Slot Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a shop for the authenticated owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopCreateInput"
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Shop slug already taken or owner already has a shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listShops
      tags: [shops]
      summary: Search/list shops
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search by name or city
        - name: city
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShopSummary"
  /shops/{slug}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get shop public profile
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shop profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopProfile"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shops/{slug}/status:
    patch:
      operationId: updateShopStatus
      tags: [shops]
      summary: Toggle shop open/closed or pause bookings
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopStatusResponse"
  /shops/{slug}/settings:
    patch:
      operationId: updateShopSettings
      tags: [shops]
      summary: Update shop settings (chairs, barbers, name)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopSettingsInput"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{slug}/dashboard:
    get:
      operationId: getShopDashboard
      tags: [shops]
      summary: Get dashboard summary (totals, today stats)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ── SERVICES ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
...
    ShopCreateInput:
      type: object
      required: [shopName, numChairs, numBarbers, city]
      properties:
        shopName:
          type: string
        numChairs:
          type: integer
          minimum: 1
        numBarbers:
          type: integer
          minimum: 1
        city:
          type: string
        address:
          type: string
        openTime:
          type: string
          description: "HH:MM format e.g. 09:00"
        closeTime:
          type: string
          description: "HH:MM format e.g. 20:00"
        pincode:
          type: string
        latitude:
          type: string
        longitude:
          type: string
...
    Shop:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        shopName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        city:
          type: string
        pincode:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
        interiorPhotos:
          type: array
          items:
            type: string
        numChairs:
          type: integer
        numBarbers:
          type: integer
        isOpen:
          type: boolean
        isPaused:
          type: boolean
        pausedUntil:
          type: ["string", "null"]
        openTime:
          type: string
        closeTime:
          type: string
        createdAt:
          type: string
      required: [id, slug, shopName, ownerName, phone, city, numChairs, numBarbers, isOpen, isPaused, openTime, closeTime, createdAt]
    ShopSummary:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        shopName:
          type: string
        ownerName:
          type: string
        city:
          type: string
        isOpen:
          type: boolean
        isPaused:
          type: boolean
        numChairs:
          type: integer
        servicesCount:
          type: integer
        minPrice:
          type: ["integer", "null"]
        latitude:
          type: ["string", "null"]
        longitude:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
      required: [id, slug, shopName, ownerName, city, isOpen, isPaused, numChairs, servicesCount]
    ShopProfile:
      type: object
      properties:
        shop:
          $ref: "#/components/schemas/Shop"
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
      required: [shop, services]
    ShopStatusInput:
      type: object
      properties:
        isOpen:
          type: boolean
        pauseMinutes:
          type: ["integer", "null"]
          description: "Pause bookings for N minutes (30 or 60). Pass null to unpause."
    ShopStatusResponse:
      type: object
      properties:
        isOpen:
          type: boolean
        isPaused:
          type: boolean
        pausedUntil:
          type: ["string", "null"]
      required: [isOpen, isPaused]
    ShopSettingsInput:
      type: object
      properties:
        ownerName:
          type: string
        phone:
          type: string
        shopName:
          type: string
        numChairs:
          type: integer
        numBarbers:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
        city:
          type: string
        address:
          type: string
        pincode:
          type: string
    DashboardSummary:
      type: object
      properties:
        todayBookings:
          type: integer
        todayRevenue:
          type: integer
        activeSlots:
          type: integer
        availableChairs:
          type: integer
        completedToday:
          type: integer
        noShowsToday:
          type: integer
        weeklyBookings:
          type: integer
        weeklyRevenue:
          type: integer
      required: [todayBookings, todayRevenue, activeSlots, availableChairs, completedToday, noShowsToday, weeklyBookings, weeklyRevenue]

    # Service
    Service:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        price:
          type: integer
          description: Price in INR
        durationMinutes:
          type: integer
          description: Service duration in minutes (buffer added server-side)
        isActive:
          type: boolean
      required: [id, shopId, name, price, durationMinutes, isActive]
...