# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Barber Slot Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  todayBookings: number;
  todayRevenue: number;
  activeSlots: number;
  availableChairs: number;
  completedToday: number;
  noShowsToday: number;
  weeklyBookings: number;
  weeklyRevenue: number;
}