# lib/api-zod/src/generated/types/shopCreateInput.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Barber Slot Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ShopCreateInput {
  shopName: string;
  /** @minimum 1 */
  numChairs: number;
  /** @minimum 1 */
  numBarbers: number;
  city: string;
  address?: string;
  /** HH:MM format e.g. 09:00 */
  openTime?: string;
  /** HH:MM format e.g. 20:00 */
  closeTime?: string;
  pincode?: string;
  latitude?: string;
  longitude?: string;
}