# lib/api-zod/src/generated/types/shopProfile.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Barber Slot Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Service } from './service';
import type { Shop } from './shop';

export interface ShopProfile {
  shop: Shop;
  services: Service[];
}